Use code from #266 to make gridder API docs
Use the autofunction directive to create doc pages for each function in
gridder and list them in the api/gridder.rst file. This is how the API
docs will be laid out later (as proposed in #266). This adds the
configuration needed to implement this for other modules and packages.
We can then make the transition smoother.

Cleaned up some docstrings while at this.